+++ src/discogs/discogs-csv.js
@@ -1,9 +1,6 @@
 import { keywordsToDiscogsGenres, keywordsToDiscogsStyles } from "../bandcamp/bandcamp.js";
 import { capitalizeEachWord } from "../modules/utils.js";
 import { Release, Track } from "../app/release.js";
-import { generateSubmissionNotes } from "./discogs.js";
-import config from "../config.js";
-import { getExtensionManifest } from "../modules/chrome.js";
 
 /**
  * Represents a Discogs CSV entry.
@@ -56,12 +53,6 @@ export class DiscogsCsv {
    * @return {DiscogsCsv} - The converted DiscogsCsv instance.
    */
   static fromRelease(release) {
-    const manifest = getExtensionManifest();
-    let b2dData = {
-      version: manifest.version,
-      submissionNotes: generateSubmissionNotes(release)
-    };
-
     return new DiscogsCsv({
       artist: release.artist,
       title: release.title,
@@ -71,7 +62,7 @@ export class DiscogsCsv {
       genres: keywordsToDiscogsGenres(release.keywords),
       styles: keywordsToDiscogsStyles(release.keywords),
       tracks: release.tracks,
-      notes: JSON.stringify(b2dData),
+      notes: JSON.stringify(release.toMetadata()),
       date: release.date.toISOString().split('T')[0],
       images: release.coverSrc.big
     });